Marker size prop is now a float

+++ src/shaders/markers.vert
@@ -4,7 +4,7 @@
 
 layout (location = 0) in vec3 pos;
 layout (location = 1) in vec4 color;
-layout (location = 2) in uint size;
+layout (location = 2) in float size;
 layout (location = 3) in uint marker;
 layout (location = 4) in float angle;
